+++ src/seednap/steps/taxonomic_assignment/__init__.py
@@ -10,7 +10,6 @@ from seednap.steps.taxonomic_assignment.blast_runner import (
     BlastTaxonomicAssigner,
 )
 from seednap.steps.taxonomic_assignment.dada2_taxonomy_runner import Dada2TaxonomyError, Dada2TaxonomyRunner
-from seednap.steps.taxonomic_assignment.decipher_runner import DecipherError, DecipherRunner
 from seednap.steps.taxonomic_assignment.ecotag_runner import EcotagError, EcotagRunner
 
 __all__ = [
@@ -24,8 +23,6 @@ __all__ = [
     "EcotagError",
     "Dada2TaxonomyRunner",
     "Dada2TaxonomyError",
-    "DecipherRunner",
-    "DecipherError",
     "TaxonomicAssigner",
     "TaxonomyMethod",
 ]